Mutate variables in draws objects

Mutate variables in a draws object.

mutate_variables(.x, ...)

# S3 method for draws_matrix
mutate_variables(.x, ...)

# S3 method for draws_array
mutate_variables(.x, ...)

# S3 method for draws_df
mutate_variables(.x, ...)

# S3 method for draws_list
mutate_variables(.x, ...)

Arguments

.x
A draws object.

...
Name-value pairs of expressions, each with length 1 or same length as in the entire input (i.e., number of iterations or draws). The name of each argument will be the name of a new variable, and the value will be its corresponding value. Use a NULL value in mutate_variables to drop a variable. New variables overwrite existing variables of the same name.

Value

Returns a draws object of the same format as .x, with variables mutated according to the expressions provided in ....

Details

In order to mutate variables in draws_matrix and draws_array objects, they are transformed to draws_df objects first and then transformed back after mutation. As those transformations are quite expensive for larger number of draws, we recommend using mutate_variables on draws_df and draws_list objects if speed is an issue.

See also

Examples

x <- as_draws_df(example_draws())
x <- subset(x, variable = c("mu", "tau"))

mutate_variables(x, tau2 = tau^2)

#> # A draws_df: 100 iterations, 4 chains, and 3 variables
#>      mu tau .chain
#> 1  2.01 2.8      1
#> 2  1.46 7.0      1
#> 3  5.81 9.7      1
#> 4  6.85 4.8      1
#> 5  1.81 2.8      1
#> 6  3.84 4.1      1
#> 7  5.47 4.0      1
#> 8  1.20 1.5      1
#> 9  0.15 3.9      1
#> 10 7.17 1.8      1
#> # ... with 390 more draws
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

mutate_variables(x, scale = 1.96 * tau, lower = mu - scale)

#> # A draws_df: 100 iterations, 4 chains, and 4 variables
#>      mu tau .chain .iteration
#> 1  2.01 2.8      1          1
#> 2  1.46 7.0      1          2
#> 3  5.81 9.7      1          3
#> 4  6.85 4.8      1          4
#> 5  1.81 2.8      1          5
#> 6  3.84 4.1      1          6
#> 7  5.47 4.0      1          7
#> 8  1.20 1.5      1          8
#> 9  0.15 3.9      1          9
#> 10 7.17 1.8      1         10
#> # ... with 390 more draws
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}