Order draws objects

Order draws objects according to iteration and chain number. By default, draws objects are ordered but subsetting or extracting parts of them may leave them in an unordered state.

order_draws(x, ...)

# S3 method for draws_matrix
order_draws(x, ...)

# S3 method for draws_array
order_draws(x, ...)

# S3 method for draws_df
order_draws(x, ...)

# S3 method for draws_list
order_draws(x, ...)

Arguments

x
An R object for which the methods are defined.

...
Arguments passed to individual methods (if applicable).

Examples

x <- as_draws_array(example_draws())
# manually select some iterations and chains
(x <- x[10:5, 4:3, ])

#> # A draws_array: 6 iterations, 2 chains, and 10 variables
#> , , variable = mu
#> 
#>          chain
#> iteration     4   3
#>        10  3.81 8.9
#>        9   0.88 8.9
#>        8   4.90 3.5
#>        7  11.82 3.7
#>        6   2.38 1.1
#> 
#> , , variable = tau
#> 
#>          chain
#> iteration    4    3
#>        10  2.7 5.96
#>        9  15.8 5.96
#>        8   3.1 0.52
#>        7   4.3 1.68
#>        6   2.3 2.70
#> 
#> , , variable = theta[1]
#> 
#>          chain
#> iteration     4    3
#>        10  7.52 13.1
#>        9   9.07 13.1
#>        8   1.50  3.7
#>        7  18.87  5.2
#>        6   0.59  5.0
#> 
#> , , variable = theta[2]
#> 
#>          chain
#> iteration    4   3
#>        10  4.3 4.7
#>        9  11.6 4.7
#>        8   6.1 4.1
#>        7  13.0 4.1
#>        6   1.1 4.3
#> 
#> # ... with 1 more iterations, and 6 more variables

# reorder iteration and chain indices
order_draws(x)

#> # A draws_array: 6 iterations, 2 chains, and 10 variables
#> , , variable = mu
#> 
#>          chain
#> iteration     3     4
#>         5 -0.03  5.48
#>         6  1.06  2.38
#>         7  3.67 11.82
#>         8  3.51  4.90
#>         9  8.85  0.88
#> 
#> , , variable = tau
#> 
#>          chain
#> iteration     3    4
#>         5 11.03  3.0
#>         6  2.70  2.3
#>         7  1.68  4.3
#>         8  0.52  3.1
#>         9  5.96 15.8
#> 
#> , , variable = theta[1]
#> 
#>          chain
#> iteration    3     4
#>         5  8.2 -0.91
#>         6  5.0  0.59
#>         7  5.2 18.87
#>         8  3.7  1.50
#>         9 13.1  9.07
#> 
#> , , variable = theta[2]
#> 
#>          chain
#> iteration   3    4
#>         5 3.2  3.2
#>         6 4.3  1.1
#>         7 4.1 13.0
#>         8 4.1  6.1
#>         9 4.7 11.6
#> 
#> # ... with 1 more iterations, and 6 more variables