Print draws_array objects

Pretty printing for draws_array objects.

# S3 method for draws_array
print(
  x,
  digits = 2,
  max_iterations = getOption("max_iterations", 5),
  max_chains = getOption("max_chains", 8),
  max_variables = getOption("max_variables", 4),
  ...
)

Arguments

x
An R object for which the methods are defined.

digits
Minimum number of significant digits to print.

max_iterations
Maximum number of iterations to print. Can be controlled globally via the "max_iterations" option (see options for details).

max_chains
Maximum number of chains to print. Can be controlled globally via the "max_chains" option (see options details).

max_variables
Maximum number of variables to print. Can be controlled globally via the "max_variables" option (see options for details).

...
Further arguments passed to the underlying print methods.

Value

A draws object of the same class as x.

Examples

x <- as_draws_array(example_draws())
print(x)

#> # A draws_array: 100 iterations, 4 chains, and 10 variables
#> , , variable = mu
#> 
#>          chain
#> iteration   1    2     3   4
#>         1 2.0  3.0  1.79 6.5
#>         2 1.5  8.2  5.99 9.1
#>         3 5.8 -1.2  2.56 0.2
#>         4 6.8 10.9  2.79 3.7
#>         5 1.8  9.8 -0.03 5.5
#> 
#> , , variable = tau
#> 
#>          chain
#> iteration   1    2    3   4
#>         1 2.8 2.80  8.7 3.8
#>         2 7.0 2.76  2.9 6.8
#>         3 9.7 0.57  8.4 5.3
#>         4 4.8 2.45  4.4 1.6
#>         5 2.8 2.80 11.0 3.0
#> 
#> , , variable = theta[1]
#> 
#>          chain
#> iteration     1     2    3     4
#>         1  3.96  6.26 13.3  5.78
#>         2  0.12  9.32  6.3  2.09
#>         3 21.25 -0.97 10.6 15.72
#>         4 14.70 12.45  5.4  2.69
#>         5  5.96  9.75  8.2 -0.91
#> 
#> , , variable = theta[2]
#> 
#>          chain
#> iteration      1    2   3   4
#>         1  0.271  1.0 2.1 5.0
#>         2 -0.069  9.4 7.3 8.2
#>         3 14.931 -1.2 5.7 6.0
#>         4  8.586 12.5 2.8 2.7
#>         5  1.156 11.9 3.2 3.2
#> 
#> # ... with 95 more iterations, and 6 more variables