Print draws_rvars objects

Pretty printing for draws_rvars objects.

# S3 method for draws_rvars
print(
  x,
  digits = 2,
  max_variables = getOption("posterior.max_variables", 8),
  summary = getOption("posterior.rvar_summary", "mean_sd"),
  reserved = FALSE,
  ...
)

Arguments

x
A draws object, or another R object for which the method is defined.

digits
The minimum number of significant digits to print.

max_variables
The maximum number of variables to print. Can be controlled globally via the "posterior.max_variables" option.

summary
What style of summary to display: "mean_sd" displays mean±sd, "median_mad" displays median±mad. If NULL, getOption("posterior.rvar_summary") is used (default "mean_sd).

reserved
Should reserved variables be included in the output? Defaults to FALSE. See reserved-variables for an overview of currently reserved variable names.

...
Further arguments passed to the underlying print() methods.

Value

A draws object of the same class as x.

Examples

x <- as_draws_rvars(example_draws())
print(x)

#> # A draws_rvars: 100 iterations, 4 chains, and 3 variables
#> $mu: rvar<100,4>[1] mean ± sd:
#> [1] 4.2 ± 3.4 
#> 
#> $tau: rvar<100,4>[1] mean ± sd:
#> [1] 4.2 ± 3.6 
#> 
#> $theta: rvar<100,4>[8] mean ± sd:
#> [1] 6.7 ± 6.3  5.3 ± 4.6  3.0 ± 6.8  4.9 ± 4.9  3.2 ± 5.1  4.0 ± 5.2  6.5 ± 5.3 
#> [8] 4.6 ± 5.3 
#>