Print or format a random variable

Printing and formatting methods for rvars.

# S3 method for rvar
print(x, ..., summary = NULL, digits = 2, color = TRUE)

# S3 method for rvar
format(x, ..., summary = NULL, digits = 2, color = FALSE)

# S3 method for rvar
str(object, ..., summary = NULL, vec.len = NULL)

Arguments

x, object
An rvar.

...
Further arguments passed to the underlying print methods.

summary
What style of summary to display: "mean_sd" displays mean±sd, "median_mad" displays median±mad. If NULL, getOption("posterior.rvar_summary") is used (default "mean_sd).

digits
Minimum number of significant digits to print.

color
Whether or not to use color when formatting the output. If TRUE, the pillar::style_num() functions may be used to produce strings containing control sequences to produce colored output on the terminal.

vec.len
Numeric (>= 0) indicating how many 'first few' elements are displayed of each vector. If NULL, defaults to getOption("str")$vec.len, which defaults to 4.

Value

For print() and str(), an invisible version of the input object.

For format(), a character vector of the same dimensions as x where each entry is of the form "mean±sd" or "median±mad", depending on the value of summary.

Details

print() and str() print out rvar objects by summarizing each element in the random variable with either its mean±sd or median±mad, depending on the value of summary. Both functions use the format() implementation for rvar objects under the hood, which returns a character vector in the mean±sd or median±mad form.

Examples

set.seed(5678)
x = rbind(
  cbind(rvar(rnorm(1000, 1)), rvar(rnorm(1000, 2))),
  cbind(rvar(rnorm(1000, 3)), rvar(rnorm(1000, 4)))
)

print(x)

#> rvar<1000>[2,2] mean ± sd:
#>      [,1]      [,2]     
#> [1,] 1 ± 1.01  2 ± 0.99 
#> [2,] 3 ± 1.00  4 ± 1.03 

print(x, summary = "median_mad")

#> rvar<1000>[2,2] median ± mad:
#>      [,1]        [,2]       
#> [1,] 1.0 ± 1.00  2.0 ± 0.98 
#> [2,] 3.0 ± 1.03  3.9 ± 1.02 

str(x)

#>  rvar<1000>[2,2]  1 ± 1.01  3 ± 1.00  2 ± 0.99  4 ± 1.03

format(x)

#>      [,1]       [,2]      
#> [1,] "1 ± 1.01" "2 ± 0.99"
#> [2,] "3 ± 1.00" "4 ± 1.03"