Pareto-smoothing minimum sample-size

Given Pareto-k computes the minimum sample size for reliable Pareto smoothed estimate (to have small probability of large error). See section 3.2.3 in Vehtari et al. (2024). This function is exported to be usable by other packages. For user-facing diagnostic functions, see pareto_min_ss and pareto_diags.

Usage

ps_min_ss(k, ...)

Arguments

k

pareto k value

...

unused

Value

minimum sample size

See also