Calculate R* convergence diagnostic

The rstar() function generates a measure of convergence for MCMC draws based on whether it is possible to determine the Markov chain that generated a draw with probability greater than chance. To do so, it fits a machine learning classifier to a training set of MCMC draws and evaluates its predictive accuracy on a testing set: giving the ratio of accuracy to predicting a chain uniformly at random.

rstar(
  x,
  split = TRUE,
  uncertainty = FALSE,
  method = "rf",
  hyperparameters = NULL,
  training_proportion = 0.7,
  nsimulations = 1000,
  ...
)

Arguments

x
A draws_df object or one coercible to a draws_df object.

split
Logical. If TRUE, compute estimate on split chains.

uncertainty
Logical. Indicates whether to provide a vector of R* values representing uncertainty in the calculated value (if TRUE) or a single value (if FALSE). Default is TRUE.

method
The machine learning classifier to use (must be available in the caret package). The default is "rf", which calls the random forest classifier.

hyperparameters
A named list of hyperparameter settings passed to the classifier. Default for random forest classifier (method = "rf") is mtry = floor(sqt(nvariables(x))); default for gradient-based model (method = "gbm") is interaction.depth = 3, n.trees = 50, shrinkage = 0.1, n.minobsinnode = 10.

training_proportion
Proportion of iterations used to train the classifier. Default is 0.7.

nsimulations
Number of R* values in the returned vector if uncertainty is TRUE. Default is 1000.

...
Other arguments passed to caret::train().

Value

A numeric vector of length 1 (by default) or length nsimulations (if uncertainty = TRUE).

Details

The rstar() function provides a measure of MCMC convergence based on whether it is possible to determine the chain that generated a particular draw with a probability greater than chance. To do so, it fits a machine learning classifier to a subset of the original MCMC draws (the training set) and evaluates its predictive accuracy on the remaining draws (the testing set). If predictive accuracy exceeds chance (i.e. predicting the chain that generated a draw uniformly at random), the diagnostic measure R* will be above 1, indicating that convergence has yet to occur. This statistic is recently developed, and it is currently unclear what is a reasonable threshold for diagnosing convergence.

The statistic, R*, is stochastic, meaning that each time the test is run, unless the random seed is fixed, it will generally produce a different result. To minimize the implications of this stochasticity, it is recommended to repeatedly run this function to calculate a distribution of R*; alternatively, an approximation to this distribution can be obtained by setting uncertainty = TRUE, although this approximation of uncertainty will generally have a lower mean.

By default, a random forest classifier is used (method = "rf"), which tends to perform best for target distributions of around 4 dimensions and above. For lower dimensional targets, gradient boosted models (called via method = "gbm") tend to have a higher classification accuracy. On a given MCMC sample, it is recommended to try both of these classifiers.

References

Ben Lambert, Aki Vehtari (2020) R*: A robust MCMC convergence diagnostic with uncertainty using gradient-boosted machines. arXiv preprint arXiv:2003.07900.

See also

Examples

library(caret)

#> Error in library(caret): there is no package called ‘caret’

x <- example_draws("eight_schools")
rstar(x)

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.

rstar(x, split = FALSE)

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.

rstar(x, method = "gbm")

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.

# can pass additional arguments to methods
rstar(x, method = "gbm", verbose = FALSE)

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.

# with uncertainty, returns a vector of R* values
hist(rstar(x, uncertainty = TRUE))

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.

hist(rstar(x, uncertainty = TRUE, nsimulations = 100))

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.

# can use other classification methods in caret library
rstar(x, method = "knn")

#> Error: Package 'caret' is required for 'rstar' to work. Please use library(caret) to load the package.