Summaries, expectations, and variances of random variables

Compute expectations (E() or mean()), probabilities (Pr()), medians (median()), and variances (variance()) from a random variable.

E(x, ...)

# S3 method for rvar
mean(x, ...)

Pr(x, ...)

# S3 method for default
Pr(x, ...)

# S3 method for logical
Pr(x, ...)

# S3 method for rvar
Pr(x, ...)

# S3 method for rvar
median(x, ...)

# S3 method for rvar
variance(x, ...)

Arguments

x
an rvar

...
further arguments passed to underlying functions (e.g., base::mean() or base::median()), such as na.rm.

Value

A numeric vector with the same dimensions as the given random variable, where each entry in the vector is the mean, median, or variance of the corresponding entry in x.

Details

Both E(), mean(), and Pr() take means over the draws dimension of the provided random variable. Pr() additionally checks that the provided rvar is a logical variable (hence, taking its expectation results in a probability). median() takes medians, and variance() takes variances.

For consistency, E() and Pr() are also defined for base arrays so that they can be used as summary functions in summarise_draws().

See also

Examples

set.seed(5678)
x = rdo(rnorm(4, mean = 1:4, sd = 2))

# These should all be ~= c(1, 2, 3, 4)
E(x)

#> [1] 0.9772733 1.9829572 2.9577323 4.0180367

mean(x)

#> [1] 0.9772733 1.9829572 2.9577323 4.0180367

median(x)

#> [1] 0.9742794 1.9544914 2.9235341 3.9592717

# This ...
Pr(x < 1.5)

#> [1] 0.60500 0.40425 0.23525 0.10925

# ... should be about the same as this:
pnorm(1.5, mean = 1:4, sd = 2)

#> [1] 0.5987063 0.4012937 0.2266274 0.1056498