Weight draws objects

Add weights to draws objects, with one weight per draw, for use in subsequent weighting operations. For reasons of numerical accuracy, weights are stored in the form of unnormalized log-weights (in a variable called .log_weight). See weights.draws for details how to extract weights from draws objects.

weight_draws(x, weights, ...)

# S3 method for draws_matrix
weight_draws(x, weights, log = FALSE, ...)

# S3 method for draws_array
weight_draws(x, weights, log = FALSE, ...)

# S3 method for draws_df
weight_draws(x, weights, log = FALSE, ...)

# S3 method for draws_list
weight_draws(x, weights, log = FALSE, ...)

Arguments

x
An R object for which the methods are defined.

weights
A vector of weights of length equal to the number of draws in x (as obtained by method ndraws). Weights will be internally stored on the log scale (in a variable called .log_weight) and will not be normalized, but normalized (non-log) weights can be returned via the weights.draws method later.

...
Arguments passed to individual methods (if applicable).

log
Logical. Indicates if the weights passed via weights are on the log scale already. Defaults to FALSE, that is, expecting weights to be on the standard (non-log) scale.

Value

A draws object of the same class as x.

See also

Examples

x <- example_draws()

# sample some random weights for illustration
wts <- rexp(ndraws(x))
head(wts)

#> [1] 3.30432494 0.08188675 0.02760205 0.33498668 3.10525972 0.36273557

# add weights
x <- weight_draws(x, weights = wts)

#> Loading required namespace: testthat

# extract weights
head(weights(x)) # defaults to normalized weights

#> [1] 7.632986e-03 1.891583e-04 6.376071e-05 7.738188e-04 7.173146e-03
#> [6] 8.379187e-04

head(weights(x, normalize=FALSE)) # recover original weights

#> [1] 3.30432494 0.08188675 0.02760205 0.33498668 3.10525972 0.36273557

head(weights(x, log=TRUE)) # get normalized log-weights

#> [1] -4.875276 -8.572926 -9.660373 -7.164173 -4.937411 -7.084590

# add weights which are already on the log scale
log_wts <- log(wts)
head(log_wts)

#> [1]  1.195232 -2.502418 -3.589865 -1.093665  1.133097 -1.014081

x <- weight_draws(x, weights = log_wts, log = TRUE)
# extract weights
head(weights(x))

#> [1] 7.632986e-03 1.891583e-04 6.376071e-05 7.738188e-04 7.173146e-03
#> [6] 8.379187e-04

head(weights(x, log=TRUE, normalize = FALSE)) # recover original log_wts

#> [1]  1.195232 -2.502418 -3.589865 -1.093665  1.133097 -1.014081