Extract Weights from Draws Objects

Extract weights from draws objects, with one weight per draw. See weight_draws for details how to add weights to draws objects.

# S3 method for draws
weights(object, log = FALSE, normalize = TRUE, ...)

Arguments

object
A draws object.

log
Logical. Indicates if weights should be returned on the log scale. Defaults to FALSE.

normalize
Logical. Indicates if weights should be normalized in order to sum to 1 on the standard scale. Defaults to TRUE.

...
Arguments passed to individual methods (if applicable).

Value

A vector of weights, with one weight per draw.

See also

Examples

x <- example_draws()

# sample some random weights for illustration
wts <- rexp(ndraws(x))
head(wts)

#> [1] 1.7317015 1.8865157 2.3504750 2.9266527 0.2427976 0.3479586

# add weights
x <- weight_draws(x, weights = wts)

# extract weights
head(weights(x)) # defaults to normalized weights

#> [1] 0.0038661205 0.0042117518 0.0052475669 0.0065339160 0.0005420592
#> [6] 0.0007768371

head(weights(x, normalize=FALSE)) # recover original weights

#> [1] 1.7317015 1.8865157 2.3504750 2.9266527 0.2427976 0.3479586

head(weights(x, log=TRUE)) # get normalized log-weights

#> [1] -5.555504 -5.469877 -5.249991 -5.030749 -7.520135 -7.160280

# add weights which are already on the log scale
log_wts <- log(wts)
head(log_wts)

#> [1]  0.5491044  0.6347316  0.8546174  1.0738593 -1.4155272 -1.0556717

x <- weight_draws(x, weights = log_wts, log = TRUE)
# extract weights
head(weights(x))

#> [1] 0.0038661205 0.0042117518 0.0052475669 0.0065339160 0.0005420592
#> [6] 0.0007768371

head(weights(x, log=TRUE, normalize = FALSE)) # recover original log_wts

#> [1]  0.5491044  0.6347316  0.8546174  1.0738593 -1.4155272 -1.0556717