Rename variables in draws objects

Rename variables within a draws object.

rename_variables(.x, ...)

# S3 method for draws
rename_variables(.x, ...)

Arguments

.x
A draws object.

...
One or more expressions separated by commas indicating the variables to rename. Existing variable names can be unquoted expressions, as in new_name = old_name, or character vectors, as in new_name = "old_name". All elements of non-scalar variables can be renamed at once.

Value

Returns a draws object of the same format as .x, with variables renamed according to the expressions provided in ....

See also

Examples

x <- as_draws_df(example_draws())

x

#> # A draws_df: 100 iterations, 4 chains, and 10 variables
#>      mu tau theta[1] theta[2] theta[3] theta[4] theta[5] theta[6]
#> 1  2.01 2.8     3.96    0.271    -0.74      2.1    0.923      1.7
#> 2  1.46 7.0     0.12   -0.069     0.95      7.3   -0.062     11.3
#> 3  5.81 9.7    21.25   14.931     1.83      1.4    0.531      7.2
#> 4  6.85 4.8    14.70    8.586     2.67      4.4    4.758      8.1
#> 5  1.81 2.8     5.96    1.156     3.11      2.0    0.769      4.7
#> 6  3.84 4.1     5.76    9.909    -1.00      5.3    5.889     -1.7
#> 7  5.47 4.0     4.03    4.151    10.15      6.6    3.741     -2.2
#> 8  1.20 1.5    -0.28    1.846     0.47      4.3    1.467      3.3
#> 9  0.15 3.9     1.81    0.661     0.86      4.5   -1.025      1.1
#> 10 7.17 1.8     6.08    8.102     7.68      5.6    7.106      8.5
#> # ... with 390 more draws, and 2 more variables
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

rename_variables(x, mean = mu, sigma = tau)

#> # A draws_df: 100 iterations, 4 chains, and 10 variables
#>    mean sigma theta[1] theta[2] theta[3] theta[4] theta[5] theta[6]
#> 1  2.01   2.8     3.96    0.271    -0.74      2.1    0.923      1.7
#> 2  1.46   7.0     0.12   -0.069     0.95      7.3   -0.062     11.3
#> 3  5.81   9.7    21.25   14.931     1.83      1.4    0.531      7.2
#> 4  6.85   4.8    14.70    8.586     2.67      4.4    4.758      8.1
#> 5  1.81   2.8     5.96    1.156     3.11      2.0    0.769      4.7
#> 6  3.84   4.1     5.76    9.909    -1.00      5.3    5.889     -1.7
#> 7  5.47   4.0     4.03    4.151    10.15      6.6    3.741     -2.2
#> 8  1.20   1.5    -0.28    1.846     0.47      4.3    1.467      3.3
#> 9  0.15   3.9     1.81    0.661     0.86      4.5   -1.025      1.1
#> 10 7.17   1.8     6.08    8.102     7.68      5.6    7.106      8.5
#> # ... with 390 more draws, and 2 more variables
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

rename_variables(x, b = `theta[1]`)

#> # A draws_df: 100 iterations, 4 chains, and 10 variables
#>      mu tau     b theta[2] theta[3] theta[4] theta[5] theta[6]
#> 1  2.01 2.8  3.96    0.271    -0.74      2.1    0.923      1.7
#> 2  1.46 7.0  0.12   -0.069     0.95      7.3   -0.062     11.3
#> 3  5.81 9.7 21.25   14.931     1.83      1.4    0.531      7.2
#> 4  6.85 4.8 14.70    8.586     2.67      4.4    4.758      8.1
#> 5  1.81 2.8  5.96    1.156     3.11      2.0    0.769      4.7
#> 6  3.84 4.1  5.76    9.909    -1.00      5.3    5.889     -1.7
#> 7  5.47 4.0  4.03    4.151    10.15      6.6    3.741     -2.2
#> 8  1.20 1.5 -0.28    1.846     0.47      4.3    1.467      3.3
#> 9  0.15 3.9  1.81    0.661     0.86      4.5   -1.025      1.1
#> 10 7.17 1.8  6.08    8.102     7.68      5.6    7.106      8.5
#> # ... with 390 more draws, and 2 more variables
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

rename_variables(x, b = "theta[1]")

#> # A draws_df: 100 iterations, 4 chains, and 10 variables
#>      mu tau     b theta[2] theta[3] theta[4] theta[5] theta[6]
#> 1  2.01 2.8  3.96    0.271    -0.74      2.1    0.923      1.7
#> 2  1.46 7.0  0.12   -0.069     0.95      7.3   -0.062     11.3
#> 3  5.81 9.7 21.25   14.931     1.83      1.4    0.531      7.2
#> 4  6.85 4.8 14.70    8.586     2.67      4.4    4.758      8.1
#> 5  1.81 2.8  5.96    1.156     3.11      2.0    0.769      4.7
#> 6  3.84 4.1  5.76    9.909    -1.00      5.3    5.889     -1.7
#> 7  5.47 4.0  4.03    4.151    10.15      6.6    3.741     -2.2
#> 8  1.20 1.5 -0.28    1.846     0.47      4.3    1.467      3.3
#> 9  0.15 3.9  1.81    0.661     0.86      4.5   -1.025      1.1
#> 10 7.17 1.8  6.08    8.102     7.68      5.6    7.106      8.5
#> # ... with 390 more draws, and 2 more variables
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

# rename all elements of 'theta' at once
rename_variables(x, alpha = theta)

#> # A draws_df: 100 iterations, 4 chains, and 10 variables
#>      mu tau alpha[1] alpha[2] alpha[3] alpha[4] alpha[5] alpha[6]
#> 1  2.01 2.8     3.96    0.271    -0.74      2.1    0.923      1.7
#> 2  1.46 7.0     0.12   -0.069     0.95      7.3   -0.062     11.3
#> 3  5.81 9.7    21.25   14.931     1.83      1.4    0.531      7.2
#> 4  6.85 4.8    14.70    8.586     2.67      4.4    4.758      8.1
#> 5  1.81 2.8     5.96    1.156     3.11      2.0    0.769      4.7
#> 6  3.84 4.1     5.76    9.909    -1.00      5.3    5.889     -1.7
#> 7  5.47 4.0     4.03    4.151    10.15      6.6    3.741     -2.2
#> 8  1.20 1.5    -0.28    1.846     0.47      4.3    1.467      3.3
#> 9  0.15 3.9     1.81    0.661     0.86      4.5   -1.025      1.1
#> 10 7.17 1.8     6.08    8.102     7.68      5.6    7.106      8.5
#> # ... with 390 more draws, and 2 more variables
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}