Resample draws objects

Resample draws objects according to given weights obtained, for instance, through importance sampling.

resample_draws(x, ...)

# S3 method for draws
resample_draws(x, weights = NULL, method = "stratified", ndraws = NULL, ...)

Arguments

x
An R object for which the methods are defined.

...
Arguments passed to individual methods (if applicable).

weights
A vector of positive weights of length equal to the number of draws in draws (as obtained by method ndraws). Weights will be internally normalized. If weights is NULL (the default), an attempt will be made to extract weights stored already in the draws object (via weight_draws). How exactly the weights influence the resampling depends on the method argument.

method
Name of the resampling method being applied. Possible choices are "simple" for simple random resampling with replacement, "simple_no_replace" for simple random resampling without replacement, "stratified" for stratified resampling with replacement and "deterministic" for deterministic resampling with replacement. Currently, "stratified" is the default as it has comparably low variance and bias with respect to ideal resampling. The latter would sample perfectly proportional to the weights but is not possible in practice where only a finite number of draws are available. For more details about resampling methods, see Kitagawa (1996).

ndraws
Number of draws to be returned. By default (NULL), ndraws is set internally to the total number of draws in x if sensible.

Value

A draws object of the same class as x.

Details

Upon usage of resample_draws, chains will be automatically merged due to subsetting of individual draws (see subset_draws for details).

References

Kitagawa, G., Monte Carlo Filter and Smoother for Non-Gaussian Nonlinear ' State Space Models, Journal of Computational and Graphical Statistics, 5(1):1-25, 1996.

Examples

x <- as_draws_df(example_draws())

# random weights for illustrationary purposes
w <- runif(ndraws(x), 0, 10)

# use default stratified sampling
x_rs <- resample_draws(x, weights = w)

#> Merging chains in order to subset via 'draw'.

summarise_draws(x_rs, default_summary_measures())

#> # A tibble: 10 x 7
#>    variable  mean median    sd   mad      q5   q95
#>    <chr>    <dbl>  <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 mu        4.29   4.38  3.39  3.45  -0.701  9.73
#>  2 tau       3.99   2.92  3.41  2.76   0.309 11.0 
#>  3 theta[1]  6.55   6.05  5.81  4.57  -0.982 17.3 
#>  4 theta[2]  5.04   5.03  4.82  4.35  -2.87  12.3 
#>  5 theta[3]  3.03   4.14  6.97  4.93 -10.8   11.5 
#>  6 theta[4]  4.89   5.26  4.78  4.70  -2.71  12.2 
#>  7 theta[5]  3.34   3.61  5.13  4.51  -5.93  11.2 
#>  8 theta[6]  4.23   4.34  5.01  4.99  -3.49  10.7 
#>  9 theta[7]  6.65   6.43  5.11  4.59  -1.19  14.5 
#> 10 theta[8]  4.83   4.84  4.55  4.59  -3.25  11.8 

# use simple random sampling
x_rs <- resample_draws(x, weights = w, method = "simple")

#> Merging chains in order to subset via 'draw'.

summarise_draws(x_rs, default_summary_measures())

#> # A tibble: 10 x 7
#>    variable  mean median    sd   mad      q5   q95
#>    <chr>    <dbl>  <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 mu        4.21   4.27  3.33  3.30  -0.854  9.15
#>  2 tau       3.81   2.76  3.28  2.58   0.332 10.2 
#>  3 theta[1]  6.42   5.61  5.58  4.74  -1.75  17.3 
#>  4 theta[2]  5.16   4.98  4.61  3.79  -2.78  12.3 
#>  5 theta[3]  3.36   4.16  6.55  4.60 -10.3   11.5 
#>  6 theta[4]  4.27   4.47  4.71  4.76  -4.40  11.9 
#>  7 theta[5]  3.19   3.75  4.88  4.09  -6.54  11.2 
#>  8 theta[6]  4.27   4.33  4.83  4.60  -3.49  10.7 
#>  9 theta[7]  5.76   5.13  4.54  3.88  -1.19  13.0 
#> 10 theta[8]  4.66   4.79  5.03  4.62  -3.44  12.2